requirements.txt
requests==2.*
requests_oauthlib==1.*
Jinja2==3.*
colorlog==6.*
colorama==0.*
oauthlib==3.*
PyYAML~=5.4.1
python-dateutil==2.*
kubernetes==19.15.0
click==8.0.3